core: crypto: merge aes-gcm.h and aes-gcm_ext.h
Merges aes-gcm.h and aes-gcm_ext.h and aligns types of function argument
etc.
